Refactor tests(2)
* Config-related tests placed in test_config
* Regex based snakefmt tests placed together and class renamed for clarity
