+++ tests/conftest.py
@@ -1,1 +1,10 @@
+import pytest
+from click.testing import CliRunner
+
+
+@pytest.fixture
+def cli_runner() -> CliRunner:
+    return CliRunner()
+
+
 pytest_plugins = "pytester"
